Clean component before marge
